07_transformers_and_attention/assignments/transformer_from_scratch/requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0
